made createTables fn reusable so jest can use it
